Fix the backward bug!
That was nasty! The issue is that we're treating tensors as immutable with one
owner at a time. That works great except when you do a binary op on the same
input twice. The grad gets "taken" for the op the first time so the second time
it will be 0—no good.

There are ways to hack around it, but it's easier just to use the Power of Math
to avoid the issue altogether!